Extracted a tool class to reduce duplicated logic between InMemoryIdentityService and PersistentIdentityService
